Event Fragment almost complete, only need to do save button.
Re-worked how event option entries are stored:
	- Arguments are no longer passed from fragment to fragment
	- Instead, preferences file now has "Temp" keys, as well as getters and setters for temp values
Now, every time EventFrag is started, a function checks whether any values in the temp preferences are not default.  If they are not default, then the option to change the field is grayed out.  Only once all options are greyed out (as in no default values are left in the temp keys) does the save button become available.

Removed plus icon, and replaced it with a "clear fields" TextView.  Presseing it resets all temp valures in preferences and navigates back to the main menu.